Move AI block lookups behind an app extension
The extension removes AI view types from the terminal model so the model can move into warp_terminal.
