+++ app/src/ai/predict/prompt_suggestions/mod.rs
@@ -3,6 +3,7 @@ use std::sync::LazyLock;
 use warpui::AppContext;
 use warpui::keymap::Keystroke;
 
+use crate::ai::blocklist::BlockListAiViewExt;
 use crate::terminal::TerminalModel;
 use crate::util::bindings::keybinding_name_to_keystroke;
